Refactored dungeon information to its own class
