+++ source/item_location.cpp
@@ -1,6 +1,7 @@
 #include "item_location.hpp"
 #include "spoiler_log.hpp"
 #include "settings.hpp"
+#include "dungeon.hpp"
 
 //Location definitions
 //Kokiri Forest
@@ -837,435 +838,6 @@ ItemLocation GC_ShopItem6                                 = ItemLocation::Base(0
 ItemLocation GC_ShopItem7                                 = ItemLocation::Base(0x2E, 0x36, "GC Shop Item 7",                                   {Category::cGoronCity, Category::cShop});
 ItemLocation GC_ShopItem8                                 = ItemLocation::Base(0x2E, 0x37, "GC Shop Item 8",                                   {Category::cGoronCity, Category::cShop});
 
-const std::vector<ItemLocation*> DT_Vanilla = {
-  &DekuTree_MapChest,
-  &DekuTree_CompassChest,
-  &DekuTree_CompassRoomSideChest,
-  &DekuTree_BasementChest,
-  &DekuTree_SlingshotChest,
-  &DekuTree_SlingshotRoomSideChest,
-  &DekuTree_GS_BasementBackRoom,
-  &DekuTree_GS_BasementGate,
-  &DekuTree_GS_BasementVines,
-  &DekuTree_GS_CompassRoom,
-};
-const std::vector<ItemLocation*> DT_MQ = {
-  &DekuTree_MQ_MapChest,
-  &DekuTree_MQ_CompassChest,
-  &DekuTree_MQ_SlingshotChest,
-  &DekuTree_MQ_SlingshotRoomBackChest,
-  &DekuTree_MQ_BasementChest,
-  &DekuTree_MQ_BeforeSpinningLogChest,
-  &DekuTree_MQ_AfterSpinningLowChest,
-  &DekuTree_MQ_DekuScrub,
-  &DekuTree_MQ_GS_Lobby,
-  &DekuTree_MQ_GS_CompassRoom,
-  &DekuTree_MQ_GS_BasementGravesRoom,
-  &DekuTree_MQ_GS_BasementBackRoom,
-};
-const std::vector<ItemLocation*> DC_Vanilla = {
-  &DodongosCavern_MapChest,
-  &DodongosCavern_CompassChest,
-  &DodongosCavern_BombFlowerPlatformChest,
-  &DodongosCavern_BombBagChest,
-  &DodongosCavern_EndOfBridgeChest,
-  &DodongosCavern_DekuScrubLobby,
-  &DodongosCavern_DekuScrubSideRoomNearDodongos,
-  &DodongosCavern_DekuScrubNearBombBagLeft,
-  &DodongosCavern_DekuScrubNearBombBagRight,
-  &DodongosCavern_GS_VinesAboveStairs,
-  &DodongosCavern_GS_Scarecrow,
-  &DodongosCavern_GS_AlcoveAboveStairs,
-  &DodongosCavern_GS_BackRoom,
-  &DodongosCavern_GS_SideRoomNearLowerLizalfos,
-};
-const std::vector<ItemLocation*> DC_MQ = {
-  &DodongosCavern_MQ_MapChest,
-  &DodongosCavern_MQ_BombBagChest,
-  &DodongosCavern_MQ_CompassChest,
-  &DodongosCavern_MQ_LarvaeRoomChest,
-  &DodongosCavern_MQ_TorchPuzzleRoomChest,
-  &DodongosCavern_MQ_UnderGraveChest,
-  &DodongosCavern_MQ_DekuScrubLobbyRear,
-  &DodongosCavern_MQ_DekuScrubLobbyFront,
-  &DodongosCavern_MQ_DekuScrubStaircase,
-  &DodongosCavern_MQ_DekuScrubSideRoomNearLowerLizalfos,
-  &DodongosCavern_MQ_GS_ScrubRoom,
-  &DodongosCavern_MQ_GS_SongOfTimeBlockRoom,
-  &DodongosCavern_MQ_GS_LizalfosRoom,
-  &DodongosCavern_MQ_GS_LarvaeRoom,
-  &DodongosCavern_MQ_GS_BackRoom,
-};
-const std::vector<ItemLocation*> Jabu_Vanilla = {
-  &JabuJabusBelly_MapChest,
-  &JabuJabusBelly_CompassChest,
-  &JabuJabusBelly_BoomerangChest,
-  &JabuJabusBelly_DekuScrub,
-  &JabuJabusBelly_GS_LobbyBasementLower,
-  &JabuJabusBelly_GS_LobbyBasementUpper,
-  &JabuJabusBelly_GS_NearBoss,
-  &JabuJabusBelly_GS_WaterSwitchRoom,
-};
-const std::vector<ItemLocation*> Jabu_MQ = {
-  &JabuJabusBelly_MQ_FirstRoomSideChest,
-  &JabuJabusBelly_MQ_MapChest,
-  &JabuJabusBelly_MQ_SecondRoomLowerChest,
-  &JabuJabusBelly_MQ_CompassChest,
-  &JabuJabusBelly_MQ_SecondRoomUpperChest,
-  &JabuJabusBelly_MQ_BasementNearSwitchesChest,
-  &JabuJabusBelly_MQ_BasementNearVinesChest,
-  &JabuJabusBelly_MQ_NearBossChest,
-  &JabuJabusBelly_MQ_FallingLikeLikeRoomChest,
-  &JabuJabusBelly_MQ_BoomerangRoomSmallChest,
-  &JabuJabusBelly_MQ_BoomerangChest,
-  &JabuJabusBelly_MQ_GS_TailPasaranRoom,
-  &JabuJabusBelly_MQ_GS_InvisibleEnemiesRoom,
-  &JabuJabusBelly_MQ_GS_BoomerangChestRoom,
-  &JabuJabusBelly_MQ_GS_NearBoss,
-};
-const std::vector<ItemLocation*> FoT_Vanilla = {
-  &ForestTemple_FirstRoomChest,
-  &ForestTemple_FirstStalfosChest,
-  &ForestTemple_RaisedIslandCourtyardChest,
-  &ForestTemple_MapChest,
-  &ForestTemple_WellChest,
-  &ForestTemple_FallingCeilingRoomChest,
-  &ForestTemple_EyeSwitchChest,
-  &ForestTemple_BossKeyChest,
-  &ForestTemple_FloormasterChest,
-  &ForestTemple_BowChest,
-  &ForestTemple_RedPoeChest,
-  &ForestTemple_BluePoeChest,
-  &ForestTemple_BasementChest,
-  &ForestTemple_GS_RaisedIslandCourtyard,
-  &ForestTemple_GS_FirstRoom,
-  &ForestTemple_GS_LevelIslandCourtyard,
-  &ForestTemple_GS_Lobby,
-  &ForestTemple_GS_Basement,
-};
-const std::vector<ItemLocation*> FoT_MQ = {
-  &ForestTemple_MQ_FirstRoomChest,
-  &ForestTemple_MQ_WolfosChest,
-  &ForestTemple_MQ_BowChest,
-  &ForestTemple_MQ_RaisedIslandCourtyardLowerChest,
-  &ForestTemple_MQ_RaisedIslandCourtyardUpperChest,
-  &ForestTemple_MQ_WellChest,
-  &ForestTemple_MQ_MapChest,
-  &ForestTemple_MQ_CompassChest,
-  &ForestTemple_MQ_FallingCeilingRoomChest,
-  &ForestTemple_MQ_BasementChest,
-  &ForestTemple_MQ_RedeadChest,
-  &ForestTemple_MQ_BossKeyChest,
-  &ForestTemple_MQ_GS_FirstHallway,
-  &ForestTemple_MQ_GS_BlockPushRoom,
-  &ForestTemple_MQ_GS_RaisedIslandCourtyard,
-  &ForestTemple_MQ_GS_LevelIslandCourtyard,
-  &ForestTemple_MQ_GS_Well,
-};
-const std::vector<ItemLocation*> FiT_Vanilla = {
-  &FireTemple_NearBossChest,
-  &FireTemple_FlareDancerChest,
-  &FireTemple_BossKeyChest,
-  &FireTemple_BigLavaRoomBlockedDoorChest,
-  &FireTemple_BigLavaRoomLowerOpenDoorChest,
-  &FireTemple_BoulderMazeLowerChest,
-  &FireTemple_BoulderMazeUpperChest,
-  &FireTemple_BoulderMazeSideRoomChest,
-  &FireTemple_BoulderMazeShortcutChest,
-  &FireTemple_ScarecrowChest,
-  &FireTemple_MapChest,
-  &FireTemple_CompassChest,
-  &FireTemple_HighestGoronChest,
-  &FireTemple_MegatonHammerChest,
-  &FireTemple_GS_SongOfTimeRoom,
-  &FireTemple_GS_BossKeyLoop,
-  &FireTemple_GS_BoulderMaze,
-  &FireTemple_GS_ScarecrowTop,
-  &FireTemple_GS_ScarecrowClimb,
-};
-const std::vector<ItemLocation*> FiT_MQ = {
-  &FireTemple_MQ_NearBossChest,
-  &FireTemple_MQ_MegatonHammerChest,
-  &FireTemple_MQ_CompassChest,
-  &FireTemple_MQ_LizalfosMazeLowerChest,
-  &FireTemple_MQ_LizalfosMazeUpperChest,
-  &FireTemple_MQ_ChestOnFire,
-  &FireTemple_MQ_MapRoomSideChest,
-  &FireTemple_MQ_MapChest,
-  &FireTemple_MQ_BossKeyChest,
-  &FireTemple_MQ_BigLavaRoomBlockedDoorChest,
-  &FireTemple_MQ_LizalfosMazeSideRoomChest,
-  &FireTemple_MQ_FreestandingKey,
-  &FireTemple_MQ_GS_AboveFireWallMaze,
-  &FireTemple_MQ_GS_FireWallMazeCenter,
-  &FireTemple_MQ_GS_BigLavaRoomOpenDoor,
-  &FireTemple_MQ_GS_FireWallMazeSideRoom,
-  &FireTemple_MQ_GS_SkullOnFire,
-};
-const std::vector<ItemLocation*> WaT_Vanilla = {
-  &WaterTemple_MapChest,
-  &WaterTemple_CompassChest,
-  &WaterTemple_TorchesChest,
-  &WaterTemple_DragonChest,
-  &WaterTemple_CentralBowTargetChest,
-  &WaterTemple_CentralPillarChest,
-  &WaterTemple_CrackedWallChest,
-  &WaterTemple_BossKeyChest,
-  &WaterTemple_LongshotChest,
-  &WaterTemple_RiverChest,
-  &WaterTemple_GS_BehindGate,
-  &WaterTemple_GS_FallingPlatformRoom,
-  &WaterTemple_GS_CentralPillar,
-  &WaterTemple_GS_NearBossKeyChest,
-  &WaterTemple_GS_River,
-};
-const std::vector<ItemLocation*> WaT_MQ = {
-  &WaterTemple_MQ_CentralPillarChest,
-  &WaterTemple_MQ_BossKeyChest,
-  &WaterTemple_MQ_LongshotChest,
-  &WaterTemple_MQ_CompassChest,
-  &WaterTemple_MQ_MapChest,
-  &WaterTemple_MQ_FreestandingKey,
-  &WaterTemple_MQ_GS_BeforeUpperWaterSwitch,
-  &WaterTemple_MQ_GS_FreestandingKeyArea,
-  &WaterTemple_MQ_GS_LizalfosHallway,
-  &WaterTemple_MQ_GS_River,
-  &WaterTemple_MQ_GS_TripleWallTorch,
-};
-const std::vector<ItemLocation*> SpT_Vanilla = {
-  &SpiritTemple_ChildBridgeChest,
-  &SpiritTemple_ChildEarlyTorchesChest,
-  &SpiritTemple_CompassChest,
-  &SpiritTemple_EarlyAdultRightChest,
-  &SpiritTemple_FirstMirrorLeftChest,
-  &SpiritTemple_FirstMirrorRightChest,
-  &SpiritTemple_MapChest,
-  &SpiritTemple_ChildClimbNorthChest,
-  &SpiritTemple_ChildClimbEastChest,
-  &SpiritTemple_SunBlockRoomChest,
-  &SpiritTemple_StatueRoomHandChest,
-  &SpiritTemple_StatueRoomNortheastChest,
-  &SpiritTemple_NearFourArmosChest,
-  &SpiritTemple_HallwayLeftInvisibleChest,
-  &SpiritTemple_HallwayRightInvisibleChest,
-  &SpiritTemple_BossKeyChest,
-  &SpiritTemple_TopmostChest,
-  &SpiritTemple_GS_HallAfterSunBlockRoom,
-  &SpiritTemple_GS_BoulderRoom,
-  &SpiritTemple_GS_Lobby,
-  &SpiritTemple_GS_SunOnFloorRoom,
-  &SpiritTemple_GS_MetalFence,
-};
-const std::vector<ItemLocation*> SpT_MQ = {
-  &SpiritTemple_MQ_EntranceFrontLeftChest,
-  &SpiritTemple_MQ_EntranceBackRightChest,
-  &SpiritTemple_MQ_EntranceFrontRightChest,
-  &SpiritTemple_MQ_EntranceBackLeftChest,
-  &SpiritTemple_MQ_ChildHammerSwitchChest,
-  &SpiritTemple_MQ_MapChest,
-  &SpiritTemple_MQ_MapRoomEnemyChest,
-  &SpiritTemple_MQ_ChildClimbNorthChest,
-  &SpiritTemple_MQ_ChildClimbSouthChest,
-  &SpiritTemple_MQ_CompassChest,
-  &SpiritTemple_MQ_StatueRoomLullabyChest,
-  &SpiritTemple_MQ_StatueRoomInvisibleChest,
-  &SpiritTemple_MQ_SilverBlockHallwayChest,
-  &SpiritTemple_MQ_SunBlockRoomChest,
-  &SpiritTemple_MQ_SymphonyRoomChest,
-  &SpiritTemple_MQ_LeeverRoomChest,
-  &SpiritTemple_MQ_BeamosRoomChest,
-  &SpiritTemple_MQ_ChestSwitchChest,
-  &SpiritTemple_MQ_BossKeyChest,
-  &SpiritTemple_MQ_MirrorPuzzleInvisibleChest,
-  &SpiritTemple_MQ_GS_SymphonyRoom,
-  &SpiritTemple_MQ_GS_LeeverRoom,
-  &SpiritTemple_MQ_GS_NineThronesRoomWest,
-  &SpiritTemple_MQ_GS_NineThronesRoomNorth,
-  &SpiritTemple_MQ_GS_SunBlockRoom,
-};
-const std::vector<ItemLocation*> ShT_Vanilla = {
-  &ShadowTemple_MapChest,
-  &ShadowTemple_HoverBootsChest,
-  &ShadowTemple_CompassChest,
-  &ShadowTemple_EarlySilverRupeeChest,
-  &ShadowTemple_InvisibleBladesVisibleChest,
-  &ShadowTemple_InvisibleBladesInvisibleChest,
-  &ShadowTemple_FallingSpikesLowerChest,
-  &ShadowTemple_FallingSpikesUpperChest,
-  &ShadowTemple_FallingSpikesSwitchChest,
-  &ShadowTemple_InvisibleSpikesChest,
-  &ShadowTemple_WindHintChest,
-  &ShadowTemple_AfterWindEnemyChest,
-  &ShadowTemple_AfterWindHiddenChest,
-  &ShadowTemple_SpikeWallsLeftChest,
-  &ShadowTemple_BossKeyChest,
-  &ShadowTemple_InvisibleFloormasterChest,
-  &ShadowTemple_FreestandingKey,
-  &ShadowTemple_GS_SingleGiantPot,
-  &ShadowTemple_GS_FallingSpikesRoom,
-  &ShadowTemple_GS_TripleGiantPot,
-  &ShadowTemple_GS_LikeLikeRoom,
-  &ShadowTemple_GS_NearShip,
-};
-const std::vector<ItemLocation*> ShT_MQ = {
-  &ShadowTemple_MQ_CompassChest,
-  &ShadowTemple_MQ_HoverBootsChest,
-  &ShadowTemple_MQ_EarlyGibdosChest,
-  &ShadowTemple_MQ_MapChest,
-  &ShadowTemple_MQ_BeamosSilverRupeesChest,
-  &ShadowTemple_MQ_FallingSpikesSwitchChest,
-  &ShadowTemple_MQ_FallingSpikesLowerChest,
-  &ShadowTemple_MQ_FallingSpikesUpperChest,
-  &ShadowTemple_MQ_InvisibleSpikesChest,
-  &ShadowTemple_MQ_BossKeyChest,
-  &ShadowTemple_MQ_SpikeWallsLeftChest,
-  &ShadowTemple_MQ_StalfosRoomChest,
-  &ShadowTemple_MQ_InvisibleBladesInvisibleChest,
-  &ShadowTemple_MQ_InvisibleBladesVisibleChest,
-  &ShadowTemple_MQ_BombFlowerChest,
-  &ShadowTemple_MQ_WindHintChest,
-  &ShadowTemple_MQ_AfterWindHiddenChest,
-  &ShadowTemple_MQ_AfterWindEnemyChest,
-  &ShadowTemple_MQ_NearShipInvisibleChest,
-  &ShadowTemple_MQ_FreestandingKey,
-  &ShadowTemple_MQ_GS_FallingSpikesRoom,
-  &ShadowTemple_MQ_GS_WindHintRoom,
-  &ShadowTemple_MQ_GS_AfterWind,
-  &ShadowTemple_MQ_GS_AfterShip,
-  &ShadowTemple_MQ_GS_NearBoss,
-};
-const std::vector<ItemLocation*> BotW_Vanilla = {
-  &BottomOfTheWell_FrontLeftFakeWallChest,
-  &BottomOfTheWell_FrontCenterBombableChest,
-  &BottomOfTheWell_RightBottomFakeWallChest,
-  &BottomOfTheWell_CompassChest,
-  &BottomOfTheWell_CenterSkulltulaChest,
-  &BottomOfTheWell_BackLeftBombableChest,
-  &BottomOfTheWell_LensOfTruthChest,
-  &BottomOfTheWell_InvisibleChest,
-  &BottomOfTheWell_UnderwaterFrontChest,
-  &BottomOfTheWell_UnderwaterLeftChest,
-  &BottomOfTheWell_MapChest,
-  &BottomOfTheWell_FireKeeseChest,
-  &BottomOfTheWell_LikeLikeChest,
-  &BottomOfTheWell_FreestandingKey,
-  &BottomOfTheWell_GS_LikeLikeCage,
-  &BottomOfTheWell_GS_EastInnerRoom,
-  &BottomOfTheWell_GS_WestInnerRoom,
-};
-const std::vector<ItemLocation*> BotW_MQ = {
-  &BottomOfTheWell_MQ_MapChest,
-  &BottomOfTheWell_MQ_LensOfTruthChest,
-  &BottomOfTheWell_MQ_CompassChest,
-  &BottomOfTheWell_MQ_DeadHandFreestandingKey,
-  &BottomOfTheWell_MQ_EastInnerRoomFreestandingKey,
-  &BottomOfTheWell_MQ_GS_Basement,
-  &BottomOfTheWell_MQ_GS_CoffinRoom,
-  &BottomOfTheWell_MQ_GS_WestInnerRoom,
-};
-const std::vector<ItemLocation*> IC_Vanilla = {
-  &IceCavern_MapChest,
-  &IceCavern_CompassChest,
-  &IceCavern_IronBootsChest,
-  &IceCavern_FreestandingPoH,
-  &IceCavern_GS_PushBlockRoom,
-  &IceCavern_GS_SpinningScytheRoom,
-  &IceCavern_GS_HeartPieceRoom,
-};
-const std::vector<ItemLocation*> IC_MQ = {
-  &IceCavern_MQ_IronBootsChest,
-  &IceCavern_MQ_CompassChest,
-  &IceCavern_MQ_MapChest,
-  &IceCavern_MQ_FreestandingPoH,
-  &IceCavern_MQ_GS_Scarecrow,
-  &IceCavern_MQ_GS_IceBlock,
-  &IceCavern_MQ_GS_RedIce,
-};
-const std::vector<ItemLocation*> GTG_Vanilla = {
-  &GerudoTrainingGrounds_LobbyLeftChest,
-  &GerudoTrainingGrounds_LobbyRightChest,
-  &GerudoTrainingGrounds_StalfosChest,
-  &GerudoTrainingGrounds_BeamosChest,
-  &GerudoTrainingGrounds_HiddenCeilingChest,
-  &GerudoTrainingGrounds_MazePathFirstChest,
-  &GerudoTrainingGrounds_MazePathSecondChest,
-  &GerudoTrainingGrounds_MazePathThirdChest,
-  &GerudoTrainingGrounds_MazePathFinalChest,
-  &GerudoTrainingGrounds_MazeRightCentralChest,
-  &GerudoTrainingGrounds_MazeRightSideChest,
-  &GerudoTrainingGrounds_UnderwaterSilverRupeeChest,
-  &GerudoTrainingGrounds_HammerRoomClearChest,
-  &GerudoTrainingGrounds_HammerRoomSwitchChest,
-  &GerudoTrainingGrounds_EyeStatueChest,
-  &GerudoTrainingGrounds_NearScarecrowChest,
-  &GerudoTrainingGrounds_BeforeHeavyBlockChest,
-  &GerudoTrainingGrounds_HeavyBlockFirstChest,
-  &GerudoTrainingGrounds_HeavyBlockSecondChest,
-  &GerudoTrainingGrounds_HeavyBlockThirdChest,
-  &GerudoTrainingGrounds_HeavyBlockFourthChest,
-  &GerudoTrainingGrounds_FreestandingKey,
-};
-const std::vector<ItemLocation*> GTG_MQ = {
-  &GerudoTrainingGrounds_MQ_LobbyRightChest,
-  &GerudoTrainingGrounds_MQ_LobbyLeftChest,
-  &GerudoTrainingGrounds_MQ_FirstIronKnuckleChest,
-  &GerudoTrainingGrounds_MQ_BeforeHeavyBlockChest,
-  &GerudoTrainingGrounds_MQ_EyeStatueChest,
-  &GerudoTrainingGrounds_MQ_FlameCircleChest,
-  &GerudoTrainingGrounds_MQ_SecondIronKnuckleChest,
-  &GerudoTrainingGrounds_MQ_DinolfosChest,
-  &GerudoTrainingGrounds_MQ_IceArrowsChest,
-  &GerudoTrainingGrounds_MQ_MazeRightCentralChest,
-  &GerudoTrainingGrounds_MQ_MazePathFirstChest,
-  &GerudoTrainingGrounds_MQ_MazeRightSideChest,
-  &GerudoTrainingGrounds_MQ_MazePathThirdChest,
-  &GerudoTrainingGrounds_MQ_MazePathSecondChest,
-  &GerudoTrainingGrounds_MQ_HiddenCeilingChest,
-  &GerudoTrainingGrounds_MQ_UnderwaterSilverRupeeChest,
-  &GerudoTrainingGrounds_MQ_HeavyBlockChest,
-};
-const std::vector<ItemLocation*> GC_Vanilla = {
-  &GanonsCastle_ForestTrialChest,
-  &GanonsCastle_WaterTrialLeftChest,
-  &GanonsCastle_WaterTrialRightChest,
-  &GanonsCastle_ShadowTrialFrontChest,
-  &GanonsCastle_ShadowTrialGoldenGauntletsChest,
-  &GanonsCastle_SpiritTrialCrystalSwitchChest,
-  &GanonsCastle_SpiritTrialInvisibleChest,
-  &GanonsCastle_LightTrialFirstLeftChest,
-  &GanonsCastle_LightTrialSecondLeftChest,
-  &GanonsCastle_LightTrialThirdLeftChest,
-  &GanonsCastle_LightTrialFirstRightChest,
-  &GanonsCastle_LightTrialSecondRightChest,
-  &GanonsCastle_LightTrialThirdRightChest,
-  &GanonsCastle_LightTrialInvisibleEnemiesChest,
-  &GanonsCastle_LightTrialLullabyChest,
-  &GanonsCastle_DekuScrubLeft,
-  &GanonsCastle_DekuScrubCenterLeft,
-  &GanonsCastle_DekuScrubCenterRight,
-  &GanonsCastle_DekuScrubRight,
-};
-const std::vector<ItemLocation*> GC_MQ = {
-  &GanonsCastle_MQ_WaterTrialChest,
-  &GanonsCastle_MQ_ForestTrialEyeSwitchChest,
-  &GanonsCastle_MQ_ForestTrialFrozenEyeSwitchChest,
-  &GanonsCastle_MQ_LightTrialLullabyChest,
-  &GanonsCastle_MQ_ShadowTrialBombFlowerChest,
-  &GanonsCastle_MQ_ShadowTrialEyeSwitchChest,
-  &GanonsCastle_MQ_SpiritTrialGoldenGauntletsChest,
-  &GanonsCastle_MQ_SpiritTrialSunBackRightChest,
-  &GanonsCastle_MQ_SpiritTrialSunBackLeftChest,
-  &GanonsCastle_MQ_SpiritTrialSunFrontLeftChest,
-  &GanonsCastle_MQ_SpiritTrialFirstChest,
-  &GanonsCastle_MQ_SpiritTrialInvisibleChest,
-  &GanonsCastle_MQ_ForestTrialFreestandingKey,
-  &GanonsCastle_MQ_DekuScrubRight,
-  &GanonsCastle_MQ_DekuScrubCenterLeft,
-  &GanonsCastle_MQ_DekuScrubCenter,
-  &GanonsCastle_MQ_DekuScrubCenterRight,
-  &GanonsCastle_MQ_DekuScrubLeft,
-};
-
 std::vector<ItemLocation*> dungeonRewardLocations = {
   //Bosses
   &QueenGohma,
@@ -1372,6 +944,7 @@ std::vector<ItemLocation*> overworldLocations = {
   &MK_TreasureChestGameReward,
   &MK_BombchuBowlingFirstPrize,
   &MK_BombchuBowlingSecondPrize,
+  &MK_BombchuBowlingBombchus,
   &MK_LostDog,
   &MK_ShootingGalleryReward,
   &MK_10BigPoes,
@@ -1647,63 +1220,9 @@ void GenerateLocationPool() {
   AddLocation(&LinksPocket);
   AddLocations(overworldLocations);
 
-  //Deku Tree
-  AddLocations((DekuTreeDungeonMode) ? DT_MQ : DT_Vanilla);
-  AddLocation(&DekuTree_QueenGohmaHeart);
-  AddLocation(&QueenGohma);
-
-  //Dodongos Cavern
-  AddLocations((DodongosCavernDungeonMode) ? DC_MQ : DC_Vanilla);
-  AddLocation(&DodongosCavern_BossRoomChest);
-  AddLocation(&DodongosCavern_KingDodongoHeart);
-  AddLocation(&KingDodongo);
-
-  //Jabu Jabu's Belly
-  AddLocations((JabuJabusBellyDungeonMode) ? Jabu_MQ : Jabu_Vanilla);
-  AddLocation(&JabuJabusBelly_BarinadeHeart);
-  AddLocation(&Barinade);
-
-  //Forest Temple
-  AddLocations((ForestTempleDungeonMode) ? FoT_MQ : FoT_Vanilla);
-  AddLocation(&ForestTemple_PhantomGanonHeart);
-  AddLocation(&PhantomGanon);
-
-  //Fire Temple
-  AddLocations((FireTempleDungeonMode) ? FiT_MQ : FiT_Vanilla);
-  AddLocation(&FireTemple_VolvagiaHeart);
-  AddLocation(&Volvagia);
-
-  //Water Temple
-  AddLocations((WaterTempleDungeonMode) ? WaT_MQ : WaT_Vanilla);
-  AddLocation(&WaterTemple_MorphaHeart);
-  AddLocation(&Morpha);
-
-  //Spirit Temple
-  AddLocation(&SpiritTemple_SilverGauntletsChest);
-  AddLocation(&SpiritTemple_MirrorShieldChest);
-  AddLocations((SpiritTempleDungeonMode) ? SpT_MQ : SpT_Vanilla);
-  AddLocation(&SpiritTemple_TwinrovaHeart);
-  AddLocation(&Twinrova);
-
-  //Shadow Temple
-  AddLocations((ShadowTempleDungeonMode) ? ShT_MQ : ShT_Vanilla);
-  AddLocation(&ShadowTemple_BongoBongoHeart);
-  AddLocation(&BongoBongo);
-
-  //Bottom of the Well
-  AddLocations((BottomOfTheWellDungeonMode) ? BotW_MQ : BotW_Vanilla);
-
-  //Ice Cavern
-  AddLocations((IceCavernDungeonMode) ? IC_MQ : IC_Vanilla);
-  AddLocation(&SheikInIceCavern);
-
-  //Gerudo Training Grounds
-  AddLocations((GerudoTrainingGroundsDungeonMode) ? GTG_MQ : GTG_Vanilla);
-
-  //Ganons Castle
-  AddLocations((GanonsCastleDungeonMode) ? GC_MQ : GC_Vanilla);
-  AddLocation(&GanonsCastle_BossKeyChest);
-  AddLocation(&Ganon);
+  for (auto dungeon : Dungeon::dungeonList) {
+    AddLocations(dungeon.GetDungeonLocations());
+  }
 }
 
 void PlaceItemInLocation(ItemLocation* loc, Item item, bool applyEffectImmediately /*= false*/) {
@@ -1760,59 +1279,14 @@ void ItemReset() {
   itemsPlaced = 0;
 }
 
-//Fills everyPossibleLocation and creates an exclusion option for each location
+//Fills everyPossibleLocation and creates an exclusion option for each location.
 //everyPossibleLocation is used in the menu to lock/hide excluding locations
 void AddExcludedOptions() {
   AddLocations(overworldLocations, &everyPossibleLocation);
 
-  AddLocations(DT_MQ, &everyPossibleLocation);
-  AddLocations(DT_Vanilla, &everyPossibleLocation);
-  AddLocation(&DekuTree_QueenGohmaHeart, &everyPossibleLocation);
-
-  AddLocations(DC_MQ, &everyPossibleLocation);
-  AddLocations(DC_Vanilla, &everyPossibleLocation);
-  AddLocation(&DodongosCavern_BossRoomChest, &everyPossibleLocation);
-  AddLocation(&DodongosCavern_KingDodongoHeart, &everyPossibleLocation);
-
-  AddLocations(Jabu_MQ, &everyPossibleLocation);
-  AddLocations(Jabu_Vanilla, &everyPossibleLocation);
-  AddLocation(&JabuJabusBelly_BarinadeHeart, &everyPossibleLocation);
-
-  AddLocations(FoT_MQ, &everyPossibleLocation);
-  AddLocations(FoT_Vanilla, &everyPossibleLocation);
-  AddLocation(&ForestTemple_PhantomGanonHeart, &everyPossibleLocation);
-
-  AddLocations(FiT_MQ, &everyPossibleLocation);
-  AddLocations(FiT_Vanilla, &everyPossibleLocation);
-  AddLocation(&FireTemple_VolvagiaHeart, &everyPossibleLocation);
-
-  AddLocations(WaT_MQ, &everyPossibleLocation);
-  AddLocations(WaT_Vanilla, &everyPossibleLocation);
-  AddLocation(&WaterTemple_MorphaHeart, &everyPossibleLocation);
-
-  AddLocation(&SpiritTemple_SilverGauntletsChest, &everyPossibleLocation);
-  AddLocation(&SpiritTemple_MirrorShieldChest, &everyPossibleLocation);
-  AddLocations(SpT_MQ, &everyPossibleLocation);
-  AddLocations(SpT_Vanilla, &everyPossibleLocation);
-  AddLocation(&SpiritTemple_TwinrovaHeart, &everyPossibleLocation);
-
-  AddLocations(ShT_MQ, &everyPossibleLocation);
-  AddLocations(ShT_Vanilla, &everyPossibleLocation);
-  AddLocation(&ShadowTemple_BongoBongoHeart, &everyPossibleLocation);
-
-  AddLocations(BotW_MQ, &everyPossibleLocation);
-  AddLocations(BotW_Vanilla, &everyPossibleLocation);
-
-  AddLocations(IC_MQ, &everyPossibleLocation);
-  AddLocations(IC_Vanilla, &everyPossibleLocation);
-  AddLocation(&SheikInIceCavern, &everyPossibleLocation);
-
-  AddLocations(GTG_MQ, &everyPossibleLocation);
-  AddLocations(GTG_Vanilla, &everyPossibleLocation);
-
-  AddLocations(GC_MQ, &everyPossibleLocation);
-  AddLocations(GC_Vanilla, &everyPossibleLocation);
-  AddLocation(&GanonsCastle_BossKeyChest, &everyPossibleLocation);
+  for (auto dungeon : Dungeon::dungeonList) {
+    AddLocations(dungeon.GetEveryLocation(), &everyPossibleLocation);
+  }
 
   for (ItemLocation * il: everyPossibleLocation) {
     il->AddExcludeOption();
